Extract from BucketsState logic resposible for eviction
