+++ engine/flink/components/base-unbounded/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/aggregate/AggregatorFunctionWithMapState.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.engine.flink.util.transformer.aggregate
 
 import org.apache.flink.api.common.functions.{OpenContext, RuntimeContext}
-import org.apache.flink.api.common.state.{MapStateDescriptor, ValueStateDescriptor}
+import org.apache.flink.api.common.state.{MapState, MapStateDescriptor, ValueState, ValueStateDescriptor}
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.java.typeutils.ListTypeInfo
 import org.apache.flink.streaming.api.TimerService
@@ -62,6 +62,9 @@ trait AggregatorFunctionWithMapStateMixin extends AggregatorFunctionBase {
   @transient
   protected var bucketsState: BucketsState = _
 
+  @transient
+  protected var latestEvictionTimeForKey: ValueState[java.lang.Long] = _
+
   protected def initBucketsState(): Unit = {
     val mapState = getRuntimeContext.getMapState(
       new MapStateDescriptor[java.lang.Long, AnyRef](
@@ -76,10 +79,10 @@ trait AggregatorFunctionWithMapStateMixin extends AggregatorFunctionBase {
         new ListTypeInfo(TypeInformation.of(classOf[java.lang.Long]))
       )
     )
-    val latestEvictionTimeForKey = getRuntimeContext.getState[java.lang.Long](
+    bucketsState = new BucketsState(mapState, keysState)
+    latestEvictionTimeForKey = getRuntimeContext.getState[java.lang.Long](
       new ValueStateDescriptor[java.lang.Long]("timers", classOf[java.lang.Long])
     )
-    bucketsState = new BucketsState(mapState, keysState, latestEvictionTimeForKey)
   }
 
   // for extending classes purpose
@@ -148,7 +151,7 @@ trait AggregatorFunctionWithMapStateMixin extends AggregatorFunctionBase {
   // Not extracted to shared trait to avoid touching LatelyEvictableStateFunctionMixin which is used
   // by unrelated components (UnionWithMemoTransformer, TransformStateTransformer).
   protected def doMoveEvictionTime(time: Long, timeService: TimerService): Unit = {
-    val latestEvictionTimeValue = bucketsState.latestEvictionTime
+    val latestEvictionTimeValue = latestEvictionTimeForKey.value()
     val maxEvictionTime = if (latestEvictionTimeValue == null || time > latestEvictionTimeValue) {
       time
     } else {
@@ -159,15 +162,16 @@ trait AggregatorFunctionWithMapStateMixin extends AggregatorFunctionBase {
       timeService.registerEventTimeTimer(maxEvictionTime)
     }
 
-    bucketsState.updateLatestEvictionTime(maxEvictionTime)
+    latestEvictionTimeForKey.update(maxEvictionTime)
   }
 
   protected def handleOnTimer(timestamp: Long, timerService: TimerService): Unit = {
-    val latestEvictionTimeValue = bucketsState.latestEvictionTime
+    val latestEvictionTimeValue = latestEvictionTimeForKey.value()
     val noLaterEventsArrived    = latestEvictionTimeValue == timestamp
 
     if (noLaterEventsArrived) {
       bucketsState.clear()
+      latestEvictionTimeForKey.update(null)
     } else if (latestEvictionTimeValue != null) {
       evictOldBuckets(timestamp)
       timerService.registerEventTimeTimer(latestEvictionTimeValue)
@@ -177,6 +181,50 @@ trait AggregatorFunctionWithMapStateMixin extends AggregatorFunctionBase {
   private def evictOldBuckets(currentWatermark: Long): Unit = {
     val cutoff = currentWatermark - timeWindowLengthMillis + 1 - allowedOutOfOrderMs
     bucketsState.removeOlderThan(cutoff)
+  }
+
+}
+
+/**
+ * Manages bucket data for MapState-based sliding window aggregators.
+ *
+ * Maintains a separate ValueState with bucket timestamps (keys) to avoid
+ * expensive prefix scans over RocksDB MapState for key-only operations
+ * like range checks, max computation, and eviction decisions.
+ */
+class BucketsState(
+    private val mapState: MapState[java.lang.Long, AnyRef],
+    private val keysState: ValueState[java.util.List[java.lang.Long]]
+) {
+
+  def keys: java.util.List[java.lang.Long] = keysState.value() match {
+    case null => new java.util.ArrayList[java.lang.Long]()
+    case v    => v
+  }
+
+  def get(key: java.lang.Long): AnyRef =
+    mapState.get(key)
+
+  def put(key: java.lang.Long, value: AnyRef): Unit = {
+    mapState.put(key, value)
+    val currentKeys = keys
+    if (!currentKeys.contains(key)) {
+      currentKeys.add(key)
+      keysState.update(currentKeys)
+    }
+  }
+
+  def removeOlderThan(cutoff: java.lang.Long): Unit = {
+    val (toRemove, toKeep) = keys.asScala.partition(_ < cutoff)
+    if (toRemove.nonEmpty) {
+      toRemove.foreach(mapState.remove)
+      keysState.update(new java.util.ArrayList(toKeep.asJava))
+    }
+  }
+
+  def clear(): Unit = {
+    mapState.clear()
+    keysState.update(null)
   }
 
 }
